Fixed run-time assert on TizenNaCl
On TizenNaCl we were hitting a runtime assert because upscaling functions were not properly initialized. This was due to a configuration issue for NaCl. NaCl has __ARM_NEON__ defined but has WEBP_USE_NEON disabled because neon does not work properly for NaCl (according to the comment in the code). Due to this the upscaling functions were not initialized at all because it was expected to be initialized by the neon code, which is not used on NaCl.

* Added additional check for WEBP_USE_NEON before defining WEBP_NEON_OMIT_C_CODE.